Update README.md
As of April 2023, there were 2,640 known billionaires (in USD) in the world. It is helpful for researchers and governmental policymakers to understand the distribution of wealth among billionaires and its implications for economic policies, taxation, and wealth distribution. Policymakers often seek to balance the potential benefits of wealth accumulation with the need to address inequality and ensure inclusive economic growth.